requirements.txt
fakeredis
redis
rq
rq-scheduler
python-nmap
openstacksdk
requests
slackclient
pymysql
